--- components/common/header/screenheader.style.js
@@ -1,21 +1,0 @@
-import { StyleSheet } from "react-native";
-
-import { COLORS, SIZES } from "../../../constants";
-
-const styles = StyleSheet.create({
-  btnContainer: {
-    width: 40,
-    height: 40,
-    backgroundColor: COLORS.white,
-    borderRadius: SIZES.small / 1.25,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  btnImg: (dimension) => ({
-    width: dimension,
-    height: dimension,
-    borderRadius: SIZES.small / 1.25,
-  }),
-});
-
-export default styles;
